Move libjive/util into jlm/util
