+++ jlm/util/Makefile.sub
@@ -3,6 +3,7 @@
 
 
 LIBUTIL_SRC = \
+    jlm/util/callbacks.cpp \
     jlm/util/common.cpp \
 	jlm/util/Statistics.cpp \
