+++ jlm/util/callbacks.cpp
@@ -3,7 +3,7 @@
  * See COPYING for terms of redistribution.
  */
 
-#include <jive/util/callbacks.hpp>
+#include <jlm/util/callbacks.hpp>
 
 namespace jive {
