+++ jlm/util/common.cpp
@@ -11,3 +11,15 @@ error::~error()
 {}
 
 }
+
+namespace jive {
+
+compiler_error::~compiler_error() noexcept
+{
+}
+
+type_error::~type_error() noexcept
+{
+}
+
+}
