--- libjive/src/common.cpp
@@ -1,18 +1,0 @@
-/*
- * Copyright 2014 Helge Bahmann <hcb@chaoticmind.net>
- * See COPYING for terms of redistribution.
- */
-
-#include <jive/common.hpp>
-
-namespace jive {
-
-compiler_error::~compiler_error() noexcept
-{
-}
-
-type_error::~type_error() noexcept
-{
-}
-
-}
