+++ libjive/src/rvsdg/tracker.cpp
@@ -4,11 +4,8 @@
  * See COPYING for terms of redistribution.
  */
 
-#include <jive/common.hpp>
 #include <jive/rvsdg/graph.hpp>
 #include <jive/rvsdg/notifiers.hpp>
-#include <jive/rvsdg/simple-node.hpp>
-#include <jive/rvsdg/tracker.hpp>
 
 using namespace std::placeholders;
